+++ tests/testthat/test-createConfig.R
@@ -2,70 +2,13 @@ context("CreateConfig")
 
 # skip unless using windows
 
-if(ifelse(is.null(Sys.info()), F, Sys.info()['sysname'] == 'Windows')){
-  
-  test_that("createModuleConfig works with keyring", {
-    
-    shinyModulePackage <- 'shinyModulePackage'
-    moduleUiFunction <- 'moduleUiFunction'
-    moduleServerFunction <- 'moduleServerFunction'
-    moduleInfoBoxFile <- 'moduleInfoBoxFile()'
-    useKeyring <- TRUE
-    moduleIcon <- "info"
-    moduleDatabaseConnectionKeyService <- 'resultDatabaseDetails'
-    moduleDatabaseConnectionKeyUsername <-  'testing'
-    
-    setting <- createModuleConfig(
-      moduleId = 'test',
-      tabName = "Test",
-      shinyModulePackage = shinyModulePackage,
-      moduleUiFunction = moduleUiFunction,
-      moduleServerFunction = moduleServerFunction,
-      moduleDatabaseConnectionKeyService = moduleDatabaseConnectionKeyService,
-      moduleDatabaseConnectionKeyUsername = moduleDatabaseConnectionKeyUsername,
-      moduleInfoBoxFile =  moduleInfoBoxFile,
-      moduleIcon = moduleIcon,
-      resultDatabaseDetails = list(a=1),
-      useKeyring = useKeyring
-    )
-    
-    testthat::expect_equal(setting$id, 'test')
-    testthat::expect_equal(setting$tabName, 'Test')
-    testthat::expect_equal(setting$tabText, 'Test')
-    
-    testthat::expect_equal(setting$shinyModulePackage, shinyModulePackage)
-    testthat::expect_equal(setting$uiFunction, moduleUiFunction)
-    testthat::expect_equal(setting$serverFunction,moduleServerFunction)
-    testthat::expect_equal(setting$databaseConnectionKeyService,moduleDatabaseConnectionKeyService)
-    testthat::expect_equal(setting$databaseConnectionKeyUsername,moduleDatabaseConnectionKeyUsername)
-    testthat::expect_equal(setting$infoBoxFile, moduleInfoBoxFile)
-    testthat::expect_equal(setting$icon, moduleIcon)
-    testthat::expect_equal(setting$keyring,useKeyring)
-    
-    # check keyring
-    keyval <- keyring::key_get(
-      service = moduleDatabaseConnectionKeyService, 
-      username = moduleDatabaseConnectionKeyUsername, 
-    )
-    
-    testthat::expect_equal(as.character(
-      jsonlite::toJSON(
-        list(a=1)
-      )), keyval)
-    
-  })
-}
-
-test_that("createModuleConfig works with environmental vars", {
+test_that("createModuleConfig works", {
   
   shinyModulePackage <- 'shinyModulePackage'
   moduleUiFunction <- 'moduleUiFunction'
   moduleServerFunction <- 'moduleServerFunction'
   moduleInfoBoxFile <- 'moduleInfoBoxFile()'
-  useKeyring <- FALSE
   moduleIcon <- "info"
-  moduleDatabaseConnectionKeyService <- 'resultDatabaseDetails'
-  moduleDatabaseConnectionKeyUsername <-  'testing'
   
   setting <- createModuleConfig(
     moduleId = 'test',
@@ -73,12 +16,8 @@ test_that("createModuleConfig works with environmental vars", {
     shinyModulePackage = shinyModulePackage,
     moduleUiFunction = moduleUiFunction,
     moduleServerFunction = moduleServerFunction,
-    moduleDatabaseConnectionKeyService = moduleDatabaseConnectionKeyService,
-    moduleDatabaseConnectionKeyUsername = moduleDatabaseConnectionKeyUsername,
     moduleInfoBoxFile =  moduleInfoBoxFile,
-    moduleIcon = moduleIcon,
-    resultDatabaseDetails = list(a=1),
-    useKeyring = useKeyring
+    moduleIcon = moduleIcon
   )
   
   testthat::expect_equal(setting$id, 'test')
@@ -88,65 +27,63 @@ test_that("createModuleConfig works with environmental vars", {
   testthat::expect_equal(setting$shinyModulePackage, shinyModulePackage)
   testthat::expect_equal(setting$uiFunction, moduleUiFunction)
   testthat::expect_equal(setting$serverFunction,moduleServerFunction)
-  testthat::expect_equal(setting$databaseConnectionKeyService,moduleDatabaseConnectionKeyService)
-  testthat::expect_equal(setting$databaseConnectionKeyUsername,moduleDatabaseConnectionKeyUsername)
   testthat::expect_equal(setting$infoBoxFile, moduleInfoBoxFile)
   testthat::expect_equal(setting$icon, moduleIcon)
-  testthat::expect_equal(setting$keyring,useKeyring)
   
-  # check env var
-  
-  var.name <- paste0(moduleDatabaseConnectionKeyService, '_', moduleDatabaseConnectionKeyUsername)
-  keyval <- Sys.getenv(var.name)
-  
-  testthat::expect_equal(
-    as.character(jsonlite::toJSON(
-      list(a=1)
-    )), keyval)
   
 })
 
 
 test_that("check included createConfigs", {
   
-conf <- createDefaultAboutConfig(useKeyring = F)
+conf <- createDefaultAboutConfig()
 testthat::expect_equal(conf$uiFunction,"aboutViewer")
 testthat::expect_equal(conf$serverFunction,"aboutServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
 
-conf <- createDefaultCharacterizationConfig(useKeyring = F)
-testthat::expect_equal(conf$uiFunction,"descriptionViewer")
-testthat::expect_equal(conf$serverFunction,"descriptionServer")
+conf <- createDefaultCharacterizationConfig()
+testthat::expect_equal(conf$uiFunction,"characterizationViewer")
+testthat::expect_equal(conf$serverFunction,"characterizationServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
 
-conf <- createDefaultCohortDiagnosticsConfig(useKeyring = F)
+conf <- createDefaultCohortDiagnosticsConfig()
 testthat::expect_equal(conf$uiFunction,"cohortDiagnosticsView")
-testthat::expect_equal(conf$serverFunction,"cohortDiagnosticsSever")
+testthat::expect_equal(conf$serverFunction,"cohortDiagnosticsServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
 
-conf <- createDefaultCohortGeneratorConfig(useKeyring = F)
+conf <- createDefaultCohortGeneratorConfig()
 testthat::expect_equal(conf$uiFunction,"cohortGeneratorViewer")
 testthat::expect_equal(conf$serverFunction,"cohortGeneratorServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
 
-conf <- createDefaultEstimationConfig(useKeyring = F)
-testthat::expect_equal(conf$uiFunction,"estimationViewer")
-testthat::expect_equal(conf$serverFunction,"estimationServer")
+conf <- createDefaultCohortMethodConfig()
+testthat::expect_equal(conf$uiFunction,"cohortMethodViewer")
+testthat::expect_equal(conf$serverFunction,"cohortMethodServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
 
-conf <- createDefaultPredictionConfig(useKeyring = F)
-testthat::expect_equal(conf$uiFunction,"predictionViewer")
-testthat::expect_equal(conf$serverFunction,"predictionServer")
+conf <- createDefaultPredictionConfig()
+testthat::expect_equal(conf$uiFunction,"patientLevelPredictionViewer")
+testthat::expect_equal(conf$serverFunction,"patientLevelPredictionServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
 
-conf <- createDefaultSCCSConfig(useKeyring = F)
+conf <- createDefaultSccsConfig()
 testthat::expect_equal(conf$uiFunction,"sccsView")
 testthat::expect_equal(conf$serverFunction,"sccsServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
 
-conf <- createDefaultMetaConfig(useKeyring = F)
+conf <- createDefaultEvidenceSynthesisConfig()
 testthat::expect_equal(conf$uiFunction,"evidenceSynthesisViewer")
 testthat::expect_equal(conf$serverFunction,"evidenceSynthesisServer")
+testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
+
+conf <- createDefaultPhevaluatorConfig()
+testthat::expect_equal(conf$uiFunction,"phevaluatorViewer")
+testthat::expect_equal(conf$serverFunction,"phevaluatorServer")
+testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")
+
+conf <- createDefaultDatasourcesConfig()
+testthat::expect_equal(conf$uiFunction,"datasourcesViewer")
+testthat::expect_equal(conf$serverFunction,"datasourcesServer")
 testthat::expect_equal(conf$shinyModulePackage,"OhdsiShinyModules")