+++ NEWS.md
@@ -1,3 +1,15 @@
+
+ShinyAppBuilder v1.2.0
+======================
+- revised config to simplify
+- shiny modules now share resultDatabaseSettings
+- added function for default resultDatabaseSettings
+- requires OhdsiShinyModules 1.2.0
+
+ShinyAppBuilder v1.1.3
+======================
+- Made simplification to API to allow calling `createShinyApp` with connectionDetails rather than a `ConnectionHandler`
+
 ShinyAppBuilder v1.1.2
 ======================
 - added package versions to footer